pyInterface: 'getWaveDescThresFromFitResult' globally available
Move the 'getWaveDescThresFromFitResult' function from the only place
where it is currently used ('calcCovMatrixForFitResult.py') to
'_waveDescThresUtils.py' and make it available in the
'pyRootPwa.utils' namespace also for other uses.
